Better dashboards table
