+++ web-admin/src/features/dashboards/listing/DashboardsTableHeader.svelte
@@ -1,88 +1,18 @@
 <script lang="ts">
-  import { beforeNavigate } from "$app/navigation";
-  import { Search } from "@rilldata/web-common/components/search";
-  import type { Table } from "@tanstack/svelte-table";
+  import MetricsExplorerIcon from "@rilldata/web-common/components/icons/MetricsExplorerIcon.svelte";
   import { getContext } from "svelte";
-  import type { Readable } from "svelte/store";
 
-  const table = getContext("table") as Readable<Table<unknown>>;
-
-  // Search
-  let filter = "";
-
-  function filterTable(filter: string) {
-    $table.setGlobalFilter(filter);
-  }
-
-  $: filterTable(filter);
-
-  beforeNavigate(() => (filter = "")); // resets filter when changing projects
+  const table = getContext("table");
 
   // Number of dashboards
   $: numDashboards = $table.getRowModel().rows.length;
-
-  // Sort
-  // function sortByTitle() {
-  //   $table.setSorting([{ id: "title", desc: false }]);
-  // }
-
-  // function sortByName() {
-  //   $table.setSorting([{ id: "name", desc: false }]);
-  // }
-
-  // function sortByLastRefreshTime() {
-  //   $table.setSorting([{ id: "lastRefreshed", desc: true }]);
-  // }
-
-  // let openSortMenu = false;
-  // function closeSortMenu() {
-  //   openSortMenu = false;
-  // }
 </script>
 
 <thead>
   <tr>
-    <td
-      class="pl-2 pr-4 py-2 max-w-[800px] flex items-center gap-x-2 bg-slate-100"
-    >
-      <!-- Search bar -->
-      <div class="px-2 grow">
-        <Search placeholder="Search" autofocus={false} bind:value={filter} />
-      </div>
-
-      <!-- Spacer -->
-      <div class="grow" />
-
-      <!-- Number of dashboards -->
-      <span>{numDashboards} dashboard{numDashboards !== 1 ? "s" : ""}</span>
-
-      <!-- Sort button -->
-      <!-- <WithTogglableFloatingElement
-        active={openSortMenu}
-        distance={4}
-        alignment="end"
-      >
-        <Button
-          on:click={() => (openSortMenu = !openSortMenu)}
-          type="secondary"
-        >
-          <span>Sort</span>
-          <CaretDownIcon />
-        </Button>
-        <Menu
-          slot="floating-element"
-          minWidth="0px"
-          on:item-select={closeSortMenu}
-          on:click-outside={closeSortMenu}
-          on:escape={closeSortMenu}
-        >
-          <MenuItem on:select={sortByTitle}>Alphabetical by title</MenuItem>
-          <MenuItem on:select={sortByName}>Alphabetical by URL</MenuItem>
-          <MenuItem on:select={sortByLastRefreshTime}
-            >Most recently refreshed</MenuItem
-          >
-        </Menu>
-      </WithTogglableFloatingElement> -->
+    <td>
+      <MetricsExplorerIcon size={"14px"} />
+      {numDashboards} dashboard{numDashboards !== 1 ? "s" : ""}
     </td>
   </tr>
 </thead>
@@ -94,6 +24,8 @@ Rounded table corners are tricky:
 -->
 <style lang="postcss">
   thead tr td {
+    @apply pl-[17px] pr-4 py-3 max-w-[800px] flex items-center gap-x-2 bg-slate-100;
+    @apply font-semibold text-gray-500;
     @apply border-y;
   }
   thead tr td:first-child {
